+++ pkg/kql/validate.go
@@ -2,7 +2,6 @@ package kql
 
 import (
 	"github.com/opencloud-eu/opencloud/pkg/ast"
-	"github.com/opencloud-eu/opencloud/services/search/pkg/query"
 )
 
 func validateAst(a *ast.Ast) error {
@@ -10,7 +9,7 @@ func validateAst(a *ast.Ast) error {
 	case *ast.OperatorNode:
 		switch node.Value {
 		case BoolAND, BoolOR:
-			return &query.StartsWithBinaryOperatorError{Node: node}
+			return &StartsWithBinaryOperatorError{Node: node}
 		}
 	}
 	return nil
@@ -21,14 +20,14 @@ func validateGroupNode(n *ast.GroupNode) error {
 	case *ast.OperatorNode:
 		switch node.Value {
 		case BoolAND, BoolOR:
-			return &query.StartsWithBinaryOperatorError{Node: node}
+			return &StartsWithBinaryOperatorError{Node: node}
 		}
 	}
 
 	if n.Key != "" {
 		for _, node := range n.Nodes {
 			if ast.NodeKey(node) != "" {
-				return &query.NamedGroupInvalidNodesError{Node: node}
+				return &NamedGroupInvalidNodesError{Node: node}
 			}
 		}
 	}
